requirements.txt
python-dotenv
psycopg2
pymongo